+++ apps/judicial-system/backend/src/app/formatters/index.ts
@@ -34,5 +34,5 @@ export { getRequestPdfAsBuffer, getRequestPdfAsString } from './requestPdf'
 export { getRulingPdfAsBuffer, getRulingPdfAsString } from './rulingPdf'
 export { createCaseFilesRecord } from './caseFilesRecordPdf'
 export { createIndictment } from './indictmentPdf'
-export { createConfirmedIndictment } from './confirmedIndictmentPdf'
+export { createConfirmedPdf } from './confirmedPdf'
 export { createSubpoena } from './subpoenaPdf'
